requirements.txt
numpy~=2.3.1
pillow~=11.3.0
pip~=23.2.1
lxml~=6.0.0
pytz~=2025.2
pywin32~=310
pandas~=2.3.1
openpyxl~=3.1.5
python-dateutil~=2.9.0.post0
certifi~=2025.7.9
img2pdf~=0.6.1
pikepdf~=9.9.0
packaging~=25.0
urllib3~=2.5.0
six~=1.17.0
et_xmlfile~=2.0.0
idna~=3.10
Deprecated~=1.2.18
wrapt~=1.17.2
xlsxwriter~=3.2.5
pytesseract~=0.3.13
easygui~=0.98.3
PyPDF2~=3.0.1
requests~=2.32.4
dotenv~=0.9.9
python-dotenv~=1.1.1
pdf2jpg~=1.1